Change language selector page route and move public index to root
