+++ frontend/e2e/index.spec.ts-snapshots/Navigating-to-renders-the-language-chooser-page-1.txt
@@ -1,1 +1,1 @@
-<div class="m-auto w-[300px] bg-white md:w-[400px] lg:w-[500px]"><div class="p-8"><h1 class="sr-only"><span lang="en">Language selection</span><span lang="fr">Sélection de la langue</span></h1><div class="w-11/12 lg:w-8/12"><a href="https://www.canada.ca/en.html"><img class="h-8 w-auto" src="https://www.canada.ca/etc/designs/canada/wet-boew/assets/sig-blk-en.svg" alt="Symbol of the Government of Canada / Symbole du gouvernement du Canada" width="300" height="28" decoding="async"></a></div><div class="mt-9 mb-2 grid grid-cols-2 gap-8 md:mx-4 lg:mx-8"><section lang="en" class="text-center"><h2 class="sr-only">Government of Canada</h2><a lang="en" class="font-lato inline-flex items-center justify-center rounded-sm border align-middle no-underline outline-offset-4 px-5 py-3 text-base border-slate-700 bg-slate-700 text-white hover:bg-sky-800 focus:bg-sky-800 w-full" href="/en/public" data-discover="true">English</a></section><section lang="fr" class="text-center"><h2 class="sr-only">Gouvernement du Canada</h2><a lang="fr" class="font-lato inline-flex items-center justify-center rounded-sm border align-middle no-underline outline-offset-4 px-5 py-3 text-base border-slate-700 bg-slate-700 text-white hover:bg-sky-800 focus:bg-sky-800 w-full" href="/fr/public" data-discover="true">Français</a></section></div></div><div class="flex items-center justify-between gap-6 bg-gray-200 p-8"><div class="w-7/12 md:w-8/12"><a lang="en" class="text-slate-700 hover:text-blue-700 hover:underline focus:text-blue-700" href="https://www.canada.ca/en/transparency/terms.html">Terms and conditions</a><span class="text-gray-400"> • </span><a lang="fr" class="text-slate-700 hover:text-blue-700 hover:underline focus:text-blue-700" href="https://www.canada.ca/fr/transparence/avis.html">Avis</a></div><div class="w-5/12 md:w-4/12"><img src="https://www.canada.ca/etc/designs/canada/wet-boew/assets/wmms-blk.svg" alt="Symbol of the Government of Canada / Symbole du gouvernement du Canada" width="300" height="71" class="h-10 w-auto"></div></div></div>
+<div class="m-auto w-[300px] bg-white md:w-[400px] lg:w-[500px]"><div class="p-8"><h1 class="sr-only"><span lang="en">Language selection</span><span lang="fr">Sélection de la langue</span></h1><div class="w-11/12 lg:w-8/12"><a href="https://www.canada.ca/en.html"><img class="h-8 w-auto" src="https://www.canada.ca/etc/designs/canada/wet-boew/assets/sig-blk-en.svg" alt="Symbol of the Government of Canada / Symbole du gouvernement du Canada" width="300" height="28" decoding="async"></a></div><div class="mt-9 mb-2 grid grid-cols-2 gap-8 md:mx-4 lg:mx-8"><section lang="en" class="text-center"><h2 class="sr-only">Government of Canada</h2><a lang="en" class="font-lato inline-flex items-center justify-center rounded-sm border align-middle no-underline outline-offset-4 px-5 py-3 text-base border-slate-700 bg-slate-700 text-white hover:bg-sky-800 focus:bg-sky-800 w-full" href="/en" data-discover="true">English</a></section><section lang="fr" class="text-center"><h2 class="sr-only">Gouvernement du Canada</h2><a lang="fr" class="font-lato inline-flex items-center justify-center rounded-sm border align-middle no-underline outline-offset-4 px-5 py-3 text-base border-slate-700 bg-slate-700 text-white hover:bg-sky-800 focus:bg-sky-800 w-full" href="/fr" data-discover="true">Français</a></section></div></div><div class="flex items-center justify-between gap-6 bg-gray-200 p-8"><div class="w-7/12 md:w-8/12"><a lang="en" class="text-slate-700 hover:text-blue-700 hover:underline focus:text-blue-700" href="https://www.canada.ca/en/transparency/terms.html">Terms and conditions</a><span class="text-gray-400"> • </span><a lang="fr" class="text-slate-700 hover:text-blue-700 hover:underline focus:text-blue-700" href="https://www.canada.ca/fr/transparence/avis.html">Avis</a></div><div class="w-5/12 md:w-4/12"><img src="https://www.canada.ca/etc/designs/canada/wet-boew/assets/wmms-blk.svg" alt="Symbol of the Government of Canada / Symbole du gouvernement du Canada" width="300" height="71" class="h-10 w-auto"></div></div></div>
