+++ frontend/tests/.server/utils/route-utils.test.ts
@@ -6,7 +6,7 @@ import { ErrorCodes } from '~/errors/error-codes';
 describe('route-utils', () => {
   describe('i18nRedirect', () => {
     it('should throw an error if no language is found', () => {
-      expect(() => i18nRedirect('routes/public/index.tsx', '')).toThrow(
+      expect(() => i18nRedirect('routes/index.tsx', '')).toThrow(
         expect.objectContaining({
           msg: 'No language found in request',
           errorCode: ErrorCodes.NO_LANGUAGE_FOUND,
@@ -15,17 +15,17 @@ describe('route-utils', () => {
     });
 
     it('should redirect to the correct path if language is found', () => {
-      const result = i18nRedirect('routes/public/index.tsx', '/en/');
+      const result = i18nRedirect('routes/index.tsx', '/en/');
 
       expect(result.status).toBe(302);
-      expect(result.headers.get('location')).toBe('/en/public');
+      expect(result.headers.get('location')).toBe('/en');
     });
 
     it('should redirect to the correct path when search params are provided', () => {
-      const result = i18nRedirect('routes/public/index.tsx', '/en/', { search: new URLSearchParams('foo=bar=baz') });
+      const result = i18nRedirect('routes/index.tsx', '/en/', { search: new URLSearchParams('foo=bar=baz') });
 
       expect(result.status).toBe(302);
-      expect(result.headers.get('location')).toBe('/en/public?foo=bar%3Dbaz');
+      expect(result.headers.get('location')).toBe('/en?foo=bar%3Dbaz');
     });
   });
 });
