+++ frontend/tests/components/menu.test.tsx
@@ -9,16 +9,16 @@ describe('Menu', () => {
   it('should correctly render a Menu with a MenuItem when the file property is provided', () => {
     const RoutesStub = createRoutesStub([
       {
-        path: '/fr/public',
+        path: '/fr',
         Component: () => (
           <Menu>
-            <MenuItem file="routes/public/index.tsx">This is a test</MenuItem>
+            <MenuItem file="routes/index.tsx">This is a test</MenuItem>
           </Menu>
         ),
       },
     ]);
 
-    const { container } = render(<RoutesStub initialEntries={['/fr/public']} />);
+    const { container } = render(<RoutesStub initialEntries={['/fr']} />);
 
     expect(container).toMatchSnapshot('expected html');
   });
@@ -26,7 +26,7 @@ describe('Menu', () => {
   it('should correctly render a Menu with a MenuItem when the to property is provided', () => {
     const RoutesStub = createRoutesStub([
       {
-        path: '/fr/public',
+        path: '/fr',
         Component: () => (
           <Menu>
             <MenuItem to="https://example.com/">This is a test</MenuItem>
@@ -35,7 +35,7 @@ describe('Menu', () => {
       },
     ]);
 
-    const { container } = render(<RoutesStub initialEntries={['/fr/public']}></RoutesStub>);
+    const { container } = render(<RoutesStub initialEntries={['/fr']} />);
 
     expect(container).toMatchSnapshot('expected html');
   });
